chore(clippy): enforce -D warnings policy (#1064)
* chore(clippy): add configuration and targeted allows (#411) [1a]

- Add clippy.toml with too-many-arguments-threshold = 14
- Add 11 targeted #[allow(clippy::type_complexity)] at specific sites
- Add scoped #[allow(dead_code)] on vendor protobuf module


* chore(clippy): fix remaining warnings — collapsible-if, pattern-char, unnecessary-max (#411)

* ci: enforce clippy -D warnings policy (#411)

---------

+++ src/sql/executor/dml_analyzed/update.rs
@@ -232,7 +232,7 @@ impl Executor {
         // Bump mod_count for auto-ANALYZE tracking.
         if cnt > 0 {
             self.stats_cache()
-                .bump_mod_count(db_id, schema.table_id, cnt as u64);
+                .bump_mod_count(db_id, schema.table_id, cnt);
             self.maybe_enqueue_auto_analyze(db_id, schema.table_id, t);
         }
